ComarchCwiczeniaTesty/ComarchCwiczeniaTesty/CalculatorService.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace ComarchCwiczeniaTesty
{
    public class CalculatorService
    {
        public int Add(int x, int y)
        {
            return x + y;
        }

        public int Subtract(int x, int y)
        {
            return x - y;
        }

        public int Multiply(int x, int y)
        {
            return x * y;
        }
        
        public float Dividy(int x, int y)
        {
            return x / (float)y;
        }
    }
}
